Renames useUserStore to userStore
Renames the user store file and updates all imports to reflect the change.

This change improves consistency and readability across the codebase by adhering to a more concise naming convention for stores.
