+++ packages/frontend-v3/src/services/absenceService.ts
